Перенес в виджет декорацию и логику дурэйшна из domain модели
