--- lib/res/theme/app_decoration.dart
@@ -1,9 +1,0 @@
-import 'package:flutter/material.dart';
-import 'package:time_tracker/res/theme/app_colors.dart';
-
-class AppDecoration {
-  static const BoxDecoration note = BoxDecoration(
-    color: AppColors.bgGreen,
-    borderRadius: BorderRadius.all(Radius.circular(15)),
-  );
-}
