+++ lib/ui/screen/components/loading_error_widget.dart
@@ -1,5 +1,6 @@
 import 'package:flutter/material.dart';
 
+// Назвал LoadingErrorWidget, потому что название ErrorWidget уже есть в Флаттере
 class LoadingErrorWidget extends StatelessWidget {
   const LoadingErrorWidget({Key? key}) : super(key: key);
